requirements.txt
PyYAML>=6.0
numpy>=1.26
pandas>=2.2
tabulate>=0.9
websocket-client>=1.8
python-binance>=1.0.19
ccxt>=4.4
